sidekiq heroku test

+++ app/workers/photo_delete_worker.rb
@@ -2,16 +2,13 @@ require 'sidekiq-scheduler'
 
 class PhotoDeleteWorker
   include Sidekiq::Worker
-  EXP_TIME = 2.hours #expiration time
+  EXP_TIME = 5.minutes #expiration time
   
   def perform
   	rejected_photos = Photo.rejected
   	rejected_photos.each do |photo|
   	  photo_expired = photo.updated_at + EXP_TIME
-      unless DateTime.now.between?(photo.updated_at, photo_expired)
-        photo.destroy
-        PHOTO_LIKE_COUNT.remove_member(photo.id.to_s)
-      end
+      photo.destroy unless DateTime.now.between?(photo.updated_at, photo_expired)
   	end
   end
 end
